+++ app/src/main/AndroidManifest.xml
@@ -5,7 +5,7 @@
     <uses-permission android:name="android.permission.INTERNET" />
 
     <application
-        android:name=".Modulos.Account.Ctrls.SGLoginFire"
+        android:name=".Modulos.Account.Ctrls.EjercicioCristian"
         android:allowBackup="true"
         android:icon="@mipmap/ic_launcher"
         android:label="@string/app_name"
@@ -13,7 +13,7 @@
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
         <activity android:name=".Modulos.CtrlSitiosGo.MainActivity" />
-        <activity android:name=".Modulos.SGSplash.Ctrls.SGSplash">
+        <activity android:name=".Modulos.SGSplash.Ctrls.Ctrls.SGSplash">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
