+++ app/src/main/java/com/example/feliz/sitiosgo/Modulos/Account/Ctrls/EjercicioCristian.java
@@ -1,11 +1,17 @@
 package com.example.feliz.sitiosgo.Modulos.Account.Ctrls;
 
 import android.app.Application;
+import com.facebook.FacebookSdk;
+import com.facebook.appevents.AppEventsLogger;
+
 
 public class EjercicioCristian extends Application {
 
     @Override
     public void onCreate() {
         super.onCreate();
+
+        FacebookSdk.sdkInitialize(getApplicationContext());
+        AppEventsLogger.activateApp(this);
     }
 }
