+++ app/src/main/res/layout/activity_sgsplash.xml
@@ -9,7 +9,7 @@
     android:paddingBottom="16dp"
     android:paddingLeft="16dp"
     android:paddingRight="16dp"
-    tools:context="com.example.feliz.sitiosgo.Modulos.SGSplash.Ctrls.SGSplash">
+    tools:context="com.example.feliz.sitiosgo.Modulos.SGSplash.Ctrls.Ctrls.SGSplash">
 
     <ImageView
         android:layout_centerHorizontal="true"
